bot_insta/requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
customtkinter==5.2.2
darkdetect==0.8.0
decorator==4.4.2
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
greenlet==3.3.2
httplib2==0.31.2
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
instagrapi==2.3.0
moviepy==1.0.3
numpy==2.4.4
oauthlib==3.3.1
packaging==26.0
pillow==12.2.0
playwright==1.58.0
proglog==0.1.12
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pyparsing==3.3.2
PySocks==1.7.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
tiktok-uploader==1.2.0
toml==0.10.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
